Updating types and moving suggestors into their own files
